Get most tests working
Lots of little fixes to get the tests running. The big change here is
that type coercion of int immediates is now handled by having
operatorfoo overloads that take ints. match_types no longer has a
special rule for intimms. So you add two Exprs, it always promotes to
the larger type, even if one of them happens to be an intimm.  For
example:

(some_u8 + 3).type() == UInt(8)
(some_u8 + Expr(3)).type() == Int(32)

Obviously this change needs to be discussed more widely. It was done
because now that IntImm can be lots of different types, I would need a
casting rule to resolve what to do if both expressions are IntImms. It
was starting to smell wrong.

+++ test/generator/example_generator.cpp
@@ -79,7 +79,7 @@ public:
         f(x, y) = max(x, y);
         g(x, y, c) = cast(output_type, f(x, y) * c * compiletime_factor * runtime_factor);
 
-        g.bound(c, 0, channels).reorder(c, x, y).unroll(c);
+        g.bound(c, 0, (int)channels).reorder(c, x, y).unroll(c);
 
         // Note that we can use the Generator method natural_vector_size()
         // here; this produces the width of the SIMD vector being targeted
